+++ packages/react-router-dom/__tests__/components/LazyComponent.tsx
@@ -1,0 +1,5 @@
+import * as React from "react";
+
+export default function LazyComponent() {
+  return <h1>Lazy</h1>;
+}